C0002_WebApi/requirements.txt
fastapi
pydantic
typing
uvicorn
psycopg2-binary
